labels should be long

+++ scripts/parse_mace4_output.py
@@ -110,7 +110,7 @@ def get_features_and_labels(
     row_count, dim = get_additional_info(cursor)
     cursor.execute("SELECT output FROM mace_output")
     features = list()
-    labels = torch.zeros([row_count])
+    labels = torch.zeros([row_count], dtype=torch.long)
     for row_index in tqdm(range(row_count)):
         output = cursor.fetchone()[0]
         features.append(get_cube_from_output(output, dim))
